🌈 Update(#26): add object-data file
